Remove `fmtident()` fallback
Commit 2c2facf848 eliminated symbol _AST_VERSION. But I failed to notice
that all but one use was to see if libast is new enough. But one use was
to test if libast was too old to provide a `fmtident()` function. Resulting
in the legacy version provided by the ksh code being used. Remove that
legacy implementation as it isn't compatible with the libast version.