Fix invalid reuse of an unrelated type
This is another case where the caller requires a pointer to a `union Value`
(now a `struct Value`) but is instead handed a pointer to an unrelated
type. Which is invalid and only works by accident before my changing
it from a `union` to a `struct` with more meta data about how the value
was set.

Related #1150