Add missing braces
Run the code through

    clang-tidy ../src/.../*.c -fix -checks="readability-braces-around-statements" -p .
    bin/style

to add missing braces. This is something I did by hand (i.e., without
using clang-tidy) a year ago for the ksh code. I deliberately did not do
it for the libast code because we knew we would be removing big chunks
of it (e.g., vmalloc).