Introduce IS_VT() macro
A lot of the warnings from the new FETCH_VT() macro is because the code
is using it to test if a particular member has been assigned a value and
possibly whether that value is non-NULL. Which is a dangerous thing to
do. So introduce the IS_VT() macro that can be used to safely test if
the stored value is a specific type.

Also switch from using fprintf(stderr,...) to DPRINTF().