Remove `getconf` from the ksh.1 man page
The `getconf` command is the only optional (i.e., not enabled by default)
builtin mentioned in the main ksh man page. Remove that text.  We do not
want people to

a) think that just executing `getconf ...` will run the ksh builtin
(since it doesn't without first doing `builtin getconf`), and

b) we want to deprecate this command and ultimately remove it in favor
of the platform version.

Related #1118