Remove global var e_devnull
That var is not needed. The entire project assumes /dev/null is a valid
path. The e_devnull var isn't used in any messages and even if it was
it wouldn't need to be localized. It looks like this var may have been
introduced back when identical string literals would each be stored
independently in the process image.