Replace struct Value Nambfp_f bfp
Struct Value member `bfp` and its associated type `Nambfp_f` is not
only unnecessary it obscures the intent due to the explicit casts it
requires. This change is probably not optimal. It represents a baby step
to increasing the sanity involving using `struct Value` (formerly `union
Value`). Replace structure member `bfp` with `shbltinp` which eliminates
the casts between the two types.

Related #1150