Feature test: Detect num cpus
Several unit tests depend on knowing the number of online CPUs to
work correctly. If that isn't defined the test may default to a value
inappropriate for the system resulting in timeouts.

This also removes three CDT tests that depend on Vmalloc features. Those
tests will never to updated to work with the system malloc.