Rename another "hdr" structure member
While investigating one of the ASAN heap-use-after-free errors I noticed
that `struct Namarray` has a "hdr" member whose name obscures its purpose.

Everything in this change is a mechanical replacement of the word "hdr"
with "namfun". With one exception: in the `array_putval()` function which
calls `free(nfp)` that causes an ASAN heap-use-after-free error. That
specific change replaces a cast (an indicator of a potential bug) with
a reference to the "hdr", now "namfun", structure member.

This does not fix the bug identified by ASAN. It just improves the
readability of the code to help make fixing the bug easier.