Add the first ksh API unit test
The unit test is incomplete but that isn't the point of this commit. The
point is that it lays the groundwork for ksh API unit tests. In particular
that ksh unit tests kind of resemble interactive tests that use
`expect`. In as much as some API tests may produce stdout and/or stderr
output that needs to be verified in addition to checking the exit status
of the API test binary.

Resolves #1168