Normalize optget() usage in enum.c
This makes the b_enum() use of optget() match that seen in other
modules such as cd_pwd.c. It also eliminates this oclint warning about
an anti-pattern:

    src/cmd/ksh93/bltins/enum.c:286:9: avoid branching statement as last
    in loop [convention|P2]

It also eliminates two of the four uses of the `error(ERROR_USAGE | x, ...)`
pattern. Bringing those error message invocations in line with all the
other code.

Related #1153