Eliminate lint about branching statement
Oclint warns:

src/cmd/ksh93/sh/name.c:2151:17: avoid branching statement as last in
loop [convention|P2]

This one is interesting because the author of the code in question was
implementing the "prefer early return" pattern. While at the same time
implementing the anti-pattern of breaking out of the loop in an
unexpected manner.