Change output() from a macro to a function
Coverity Scan #279494 warns about a possible mistake involving a "Logical
vs. bitwise operator" situation. The code is correct and the Coverity
warning is spurious. Nonetheless, `output()` should not be a macro. There
is no reason it needs to be expanded inline for performance. So turn it
into a function which should silence the Coverity warning.