Coverity 334469 Logically dead code
It's not clear if this is a false positive or not. At the moment
`sh_fmtqf()` is only called from one place and always with `fold` == -1.
Nonetheless, in theory the function could be called with `fold` greater
than eight. This restructures the code into a more idiomatic pattern that
will hopefully silence the Coverity warning.