Convert mb1char() into an inlined function
I was pondering Coverity Scan 253730 which rightly points out that the

    for (c = type; c; c--) mb1char(v);

in function varsub() is problematic due to ignoring the returned value
in the context of incrementing `v`. So convert it to an inlined function
which will make it possible to eliminate such warnings and add some safety
and clarity.