Coverity 323922 Operands don't affect result
This is a bit more invasive than usual for this type of Coverity Scan
warning. Several lexstates macros are invoked with args that are unsigned
ints. Due to how macro expansion is handled this, rightly, leads lint tools
to conclude that the `c < 0 ? 0 : something` ternary is always false. Which
is itself a subtle source of bugs but that is a problem for another day.