Lint: missing default in switch statements
It isn't completely clear whether the original `switch() { ... }` block
enumerates all possible cases. And thus a `default: { abort(); }` is
appropriate. At first glance that appears to be the case; especially since
it does explicitly test for `SF_ECONT`. But that is not at all certain
let alone obvious. So rather than add a "default" case restructure the
code. This eliminates the need for a goto label.

FWIW, I think the existing code relies too heavily on `switch` statements
where other constructs would be simpler and clearer.