More replacements of zero with NULL
This builds on the previous two commits. At this point I believe that
nearly all the assignements of zero that should be NULL in the src/cmd
tree should be fixed.  While some of the src/lib code was cleaned up it
still contains a significant number of zero assignments that should be NULL
assignments because the focus of these changes was on the code in src/cmd.