Coverity 253557 Logically dead code
Normally I would add a regression test. But in this case I don't see the
point since ksh does not call any of the `sfscanf()` variants and probably
never will. I'm fixing this solely to resolve the Coverity warning.

Note that there isn't a single use of the `%ls` format in the existing
SFIO unit tests. Not to mention the function being fixed is never run by
any of the existing tests. Which implies that there is incomplete coverage
of SFIO handling of multibyte encodings.

Fixes #1253