Make libast API tests hermetic
The libast API tests prior to this change simply ran in the current working
directory (the build dir) with the environment they inherited. That can
result in false positive failures. For example, due to two tests that
create a file with the same name.

Note that the API/sfio/talarm test fails a lot of the time with this
change. That is because it writes the following message to stderr
without exiting with a non-zero status:

    Line=45: System call was automatically resumed by the OS

This change causes such unexpected messages to be treated as a test
failure. Whereas the test framework prior to this change silently ignores
stderr output.

This also removes the special-casing of "*.out" pathnames since it caused
me problems when implementing this feature. That was due to creating a
*.out test file to confirm that if it didn't match the stdout of an API
test the conclusion is the test failed.

Finally, this renames to CDT API tests to avoid name collisions with
SFIO API tests of the same name. This allows me to remove the hack I
implemented a few weeks ago to ensure the test executables had unique
names to work around a Meson bug.

Resolves #1243