Remove some stderr /dev/null redirections
While investigating ASAN failures I noticed some SIGABRTs that were
clearly due to ASAN detecting a problem. Yet there wasn't the usual ASAN
state dump. That is due to redirecting stderr to /dev/null. Something
the legacy unit tests do way too much of the time. Even when it is
completely pointless as in this statement:

    LANG=C > /dev/null 2>&1

This also replaces the hard to read ASAN options string with an array of
options that is converted to the requisite string for the env var.