Remove more stderr /dev/null redirections
The `namespace` unit test fails under ASAN but we don't know why because
of the stderr error redirections to /dev/null. With this change we can
see that ASAN reports a heap-buffer-overflow error in `nv_create()`.

I am seriously tempted to remove every single redirection to /dev/null
since so many of them appear to be pointless and hide bugs.