Fix a heap-buffer-overflow bug
If the length of string `xp` is greater than the length of string `cp`
the expression will access beyond the buffer holding `cp`.

Also, it is clear the condition is meant to only be true if the strings
compare equal in the first `xlen` chars but it is true if they are not
equal! Too, as best I can determine the body of that `if()` is never
executed; either with the original or corrected condition.