Don't use SIGBUS in signal tests
ASAN installs handlers for SIGBUS, SIGSEGV, and SIGABRT. Which makes
using those signals in a unit test problematic. Not to mention that in
practice (i.e., outside of a contrived unit test) a ksh script can't do
anything useful if the kernel sends those signals as opposed to a user
process.