Fix trap handling
My recent commit 9c1d9240 that replaced using SIGBUS with SIGUSR2
(typically the highest numbered signal) exposed a bug in trap handling.
As I predicted long ago when I fixed a different bug with trap management
there are two places in the code that confuse the `trapmax` member of
`struct sh_scoped` as being the highest numbered signal number rather than
the number of signals (i.e., size of `trapcom`). This is because the name
`trapmax` is inherently confusing. The `nsig++` being removed causes the
code to access an element past the end of the `trapcom` array.