Eliminate use of atexit() in unit tests
A recent change now runs API tests via a test runner that is responsible
for creating a temp dir that API tests run in and cleaning it up if the
test passes. So there is no longer a need for individual tests to delete
files when they complete. On the contrary, doing so may make it harder
to diagnose what went wrong if the test fails. So remove those atexit()
handlers.