Cache pointers already handled during DPRINT_*()
Calling one of the DPRINT_*() macros can cause pointers to nested data
structures to be followed. This could result in a closed cycle. Even if
it doesn't do so the same pointer may be seen more than once and we only
want to emit debug output the first time we see the pointer.