Workaround issue #1267 involving stale nameref var
Normally I wouldn't paper over a test failure in this fashion. But in this
case it is unlikely any real-world script will run into this problem. And
even if it does the fact it doesn't run under ASAN or Valgrind means it
is unlikely to notice the bug. And this change reduces the ASAN failures
to just one. Which takes us one step closer to being able to run CI
tests with ASAN enabled to avoid regressions.

Related #1267