ASAN heap-use-after-free in fault.c
The current implementation will attempt to dereference the string pointed
to by the `trap` var if a second instance of the signal is received while
the first instance is being handled and the trap unsets or redefines the
trap. Fix that by always ensuring we're using the current trap string
when handling a signal.

Fixes #658
Fixes #1272