Baby step to resolving issue #1038
This does not change any #define's or logic other than cosmetically. It
replaces some hex constants with bit shifts to clarify the meaning of a
symbol. For example, #define NV_UNJUST 0x800000 with (1 << 23). It moves
some #define's to make the relationship with related symbols clearer. It
also adds more symbols to the list recognized by `nvflag_to_syms()`.

Related #1038