Implement nv_isflag() to test if flags are set
There are lots of places in the code that do `flags & NV_XXX` to determine
a flag is set.  But solely for the truth value; i.e., is the flag set. So
add a nv_isflag() function to make it clearer and force the result to be
a bool. The need for this came to my attention when working on expanding
`nvflag_t` to 64 bits. The problem is that a lot of places in the code
use a `flags & NV_XXX` expression as an argument to a function which
takes an int (which itself should be a bool). Without an explicit ` ==
NV_XXXX` the arg is 64 bits which then gets truncated to 32 bits which
will lose any bits in the high word.

As part of this implement the first use of the new `nv_isflag()` function.