Fix unit test compatibility with OpenBSD
A recent change by me introduced the `/bsd` path in $PATH. That fixed
many problems running tests on Solaris. It had no effect on any other
platform than OpenBSD. That platform, however, has a file named `/bsd`.
But because it isn't a directory it causes the mechanism for deciding if
bsd or att universe behavior should be in effect.