Disable vi.exp on Solaris
I've spent several hours debugging this. Manually doing the same steps
works fine. It's only with running the unit test that some operations,
such as `^`, don't have the expected outcome. Presumably due to a quirk
of the `expect` utility on Solaris. For example, the output of `stty -a`
is different when running this test compared to the settings for an
actual interactive shell.