Simplify API test code
API unit tests are now run via a test runner framework that ensures each
API test invocation has a unique temp dir. So it is no longer necessary or
desirable for the `tstfile()` function to put temp files in TMPDIR. It's
also counter productive to have the unit test remove those files when the
test fails since the content of the files may be helpful in understanding
why the test failed.

Resolves #1278