Better cleanup of successful ksh unit tests
The existing logic fails to remove a temp test dir if the test exits
early due to compatibility issues; e.g., the test is known not to work
on platform "X". Whether or not the temp dir should be removed is the
responsibility of the test runner script. So move that logic from the
test postscript to the test runner script.

Related #1278