More cleanup of the chmod builtin
This removes the remaining label inside a switch in chmod.c. It doesn't
make any sense to try the chmod() when fts_read() returns FTS_DNR or
FTS_ERR.  I checked both the BSD and GNU source and neither attempt the
chmod() when those fts_read() statuses are returned.

Related #1146