Rename ast_sys.h => spawnvex.h
The ast_sys.h used to be dynamically generated and contained definitions
for spawnvex() and friends but also other symbols. We no longer dynamically
generate this header. And it no longer contains any non spawnvex() related
symbols. So rename it to make its purose clearer.  Unsurprisingly, this
reveals quite a few places that indirectly depended on #include's in that
header. Which is a IWYU mistake. So fix those missing #include's as well.