Fix bug introduced by commit f0bbd8b82
Oclint was issuing this warning:

    src/cmd/ksh93/sh/name.c:1027:41: The left operand of '&' is a
    garbage value

This is because when I introduced the `nvflags` var in commit f0bbd8b82
I missed one place that was abusing the `n` var to hold namval flags in
addition to serving other purposes. Surprisingly that mistake does not
cause any unit test failures. Which suggests there may be other problems
with the code.