Stop dynamically generating conftab.c and conftab.h
GCC 9.1 changes the output of `gcc -E` which breaks the conf.sh script.
We want to eliminate the `getconf` builtin and the `astconf()` function.
No unit tests depend on either of those. And it is highly unlikely any
real world scripts do so. So as the first step in removing those APIs and
to make building this possible with gcc 9.1 just use a static version of
conftab.c and conftab.h.

Fixes #1315
Related #1118