Stop calling external uname from the uname builtin
Running /usr/bin/uname when an unrecognized flag is seen is wrong in at
least two ways. First, the hardcoding of the /usr/bin/uname path. That
path should be detected at build time as is done for other external command
dependencies such as ed. Second, and far more serious, is that this builtin
has no business executing an external command of the same name. The whole
point of a builtin is to avoid exec()'ing an external command and provide a
cross-platform implementation that a ksh user/script can depend on. If the
user needs the behavior of the external command provided by their platform
they should be using it rather than the ksh builtin. This undocumented
under the hood redirection to the external command is dangerous.

Fixes #1317