Remove astconf("PATH_LEADING_SLASHES")
This feature was not actually implemented. That is, calling
astconf("PATH_LEADING_SLASHES") always returns an empty string. There
is no way to set it to "1". So these are three more uses of astconf()
that can be eliminated.

Note: Having a way to preserve two leading slashes is potentially useful
since POSIX says the meaning of that sequence is implementation defined
by the platform. Whereas three or more consecutive slashes is always
equivalent to a single slash. So the incomplete implementation of
PATH_LEADING_SLASHES didn't even get that correct.