Remove PROC_PARANOID and astconf("SH")
PROC_PARANOID isn't used by any ksh code and the probability it will
be used in the future is close to zero. This allows us to remove one
astconf("SH") call. The other one in pathshell() can be removed because it
always returns an empty string which makes the subsequent "if" condition
to always be true.

Resolves #1318