Eliminate sole use of astconfdisc()
The only time newconf() is called is to export _AST_FEATURES. From the
getconf documentation:

    "[+_AST_FEATURES?Process local writable values that are "
    "different from the default are stored in the \b_AST_FEATURES\b "
    "environment variable. The \b_AST_FEATURES\b value is a "
    "space-separated list of \aname\a \apath\a \avalue\a 3-tuples, "
    "where \aname\a is the system configuration name, \apath\a is "
    "the corresponding path, \b-\b if no path is applicable, and "
    "\avalue\a is the system configuration value. \b_AST_FEATURES\b "
    "is an implementation detail of process inheritance; it may "
    "change or vanish in the future; don't rely on it.]"

Time for it to go. With this change the only remaining direct uses of
astconf() are to fetch the UNIVERSE conf setting. There are some indirect
uses via the getconf() macro that I'll be addressing in subsequent changes.