Remove astconf() calls from ulimit builtin
There is no good reason for the `ulimit` builtin to fall back to whatever
value was stored in the conftab table via astconf(). The value returned by
getrlimit() is canonical. This removes what I hope is the last astconf()
call that is not for the purpose of determining the current "universe".