Fix off by one error in _dprintf()
I noticed that debug messages (e.g., from DPRINTF) longer than 511 chars
don't include a newline char. That's because the code I wrote has an
off by one error. Sigh! This fixes that and also makes it clearer when
the debug message has been truncated by adding `...` to the end of the
message when it is too long. This also increases the debug message
buffer from 512 to 1024 chars.