Silence some Coverity Scan warnings
Coverity Scan was warning that several places that do the equivalent of

    #define NV_JSON 0
    if (flags & NV_JSON) do_something();

will never call do_something() because the condition is never true.
Ideally the code would be refactored. This is a simpler way to silence
those warnings. As a bonus introducing NV_INVALID as a synonym for the
high bit in a nvflag_t can be used to detect if the code incorrectly
attempts to set that bit (e.g., by using ~0 as a nvflag_t value).

Fixes #1322