Change cs_path() implementation
Coverity Scan warns about memory leaks for code paths involving the
cs_path() function I introduced a week ago in commit 9755dce. That's a
false positive since the function uses the singleton pattern. Still, it's
easier to just change the definition to use a static fixed size buffer
rather than dynamically allocate the buffer the first time it is needed.
The alternative is to play whack-a-mole supressing these warnings every
time a new use of cs_path() is introduced.