Fix test harness bug
I recently introduce the `NO_SHCOMP` env var to control whether `meson
test` executes or skips the `/shcomp` test variants. That's because those
tests never fail unless the non `/shcomp` variant fails (ignoring the
blacklisted tests). So when running the tests with an expensive option
like ASAN or VALGRIND enabled it's nice to be able to easily skip the
`/shcomp` variants. However, that enhancement had a bug which causes every
`/shcomp` variant to be reported as failed. This fixes that.