Revert two substitutions of memmove()
When I started contributing changes the first thing I had to do was get
it to build and pass most unit tests on macOS. MacOS helpfully reports
when calling strcpy() or memcpy() involves overlapping buffers since that
is a situation with undefined behavior. Not knowing if the buffers in
those situations could legitimately overlap I simply replaced all such
function calls with memmove(). Fast forward to now.

We have fixed a lot of off by one errors in the past two years that could
result in functions doing overlapping copies where doing so is unsafe.
This reverts all but one such change (that one being in the path.c module).

Related #676