Fix Broken time And times Built-ins
Currently, the `time` and `times` built-ins result in an assertion failure
causing a crash. This is due to an erroneous assertion from commit
595f499709319d28ebbdbedcd69b2bff223e6f3a which attempted to fix a linter
warning about the usage of an uninitialized `timeval` variable `tb`.

This change removes the assertion and instead runs `memset` on `tb` in the case
where it would have been used without having been initialized. This fixes the
`cppcheck` linter warning and restores the `time` and `times` built-ins to
working order.

A simple set of tests is added to ensure that these built-ins do not break in
the future.