Rename ulimit.h to avoid conflicts with sys hdr
I ran IWYU again and it complained that a lot of symbols defined in the
project's ulimit.h header were not defined. That's because it was using
the system header of the same name. With some work we could probably get
that tool to distinguish between the two headers. However, it is easier
to just rename the project header.