Fix `TIMEFORMAT=%3P`
The old code fails to produce the correct percent utilization if the
precision was anything other than two. It was also badly structured.
This fixes the former and only partially addresses the latter. TBD is
a more comprehensive refactoring that adds support for `getrusage()`
for more fine grained times.

Fixes #1333