Remove unused `outtype()` function
The `outtype()` function was only used if the preprocessor symbol
SHOPT_FIXEDARRAY was true. Since that symbol was not enabled by default
and the code contained bugs it was removed. The only reason `outtype()`
was not removed at the same time is that lint tools failed to notice it
wasn't actually used due to the recursive invocation.