Eliminate cached shp->gd->lim.clk_tck
Caching the return value of `sysconf(_SC_CLK_TCK)` cannot be justified.
The value is used by two commands: `time` and `times`. Even then it is
only used if the platform does not provide `getrusage()`. Which is not
true for any platform we currently try to support. But even if we do have
to fallback to the `times()` API the cost of the `sysconf(_SC_CLK_TCK)`
call each time either command is used is inconsequential.