Establish new ksh unit test naming idiom
It would be nice if you could tell from just the test file name whether
the primary purpose is to test the behavior of a builtin command or more
abstract concepts. This does that by renaming all the builtin command
unit tests to have a "b_" prefix which is the idiom in the source code.

Related #1344