Remove modei()
The modei() function is a no-op as used so remove it to simplify the code
and improve test coverage. Note that it used to be used as a consequence
of programs like `pax` that used to be part of this project but have been
removed to keep the focus on ksh.

Related #1252