Remove math function `fpclassify()`.
The math function `fpclassify()` is not portable since the values it
returns are not portable without exposing C symbols like `FP_NAN` from
the math.h header file.

This is technically a non-backward compatible change that should only
be done in a major release. However, the `fpclassify()` is unusable as
currently implemented; at least in any portable fashion. The probability
that any ksh script depends on `fpclassify()` is zero, or so close to
it, that removing it is a non-issue. Therefore removing this "feature"
should be harmless.

Fixes #1346