Remove pointless `default:` switch cases
Test code coverage points out that many `default:` cases of a `optget()`
loop can never be executed. At least if the `optget()` function isn't
broken. That's because in all of these cases either a known flag or the
`:` or `?` cases will be true. Which means the `default:` case can never
be true. This improves test coverage while simultaneously making it clear,
in a highly stylized fashion, that there isn't a need for a default block.

Note that there are some commands, such as `chmod`, where the above is
not true. Since they explicitly allow arguments that superficially look
like flags by being preceded by a `-` char.