Remove an undocumented feature of `test`
What exactly is the purpose of the code I'm removing? The innermost
block being removed is executed by constructs like this:

    test -? --

Which isn't legal but I fail to see why it is special cased. It appears
to be an undocumented means of displaying a basic "usage" statement that
is so devoid of useful information as to be useless. No other `test`
implementation I have access to does anything special for that pair of
tokens.