Simplify base64 implementation
There is a lot of lines and branches in base64.c that are not exercised
by the unit tests. Almost all of them involve calling base64encode()
with NULL for the output buffer. Something that none of the existing code
does. So remove that feature.

Too, the code was needlessly confusing because it used some vars, such
as `n`, for multiple purposes.

Finally, it is BROKEN if you pass NULL for the result buffer. That's
because it then returns a pointer to the stack which cannot be safely
used when base64encode() returns.