Add some fmtmatch() tests
This is meant to mirror commit 32dfd7612 but for the fmtmatch()
function. This in no way verifies that the resulting ksh patterns are valid
or equivalent to the ERE pattern. It simply codifies, in a unit test,
the current behavior. So that if future changes inadvertently alter the
behavior of fmtmatch() the test will hopefully catch the change in behavior
and force the author to either update the test or correct their change.