Try to decrease random test timeouts
At least some of the random test timeouts are because the `read -t0.01`
in the `empty_fifos` function hangs. It looks like once in a while there
is a race that causes the SIGALRM used to abort the read to be dropped
on the floor.

(cherry picked from commit 6d38bb5e4f59dba3f01bffc516b85405c3f36e6a)