Clarify structure management in subshell.c
There is much more to be done in this area. This is the bare mininum to
disentangle the allocation of `struct Link` and `Namval_t` structures in
`sh_assignok()` and downstream code. This makes the logic a lot more
obvious at the cost of an additional malloc()/free() pair of function
calls.

Related #1002
Fixes #1375