Move the `whence` documentation into a text file
This is the very first step in resolving issue #507. It moves the text for
the documentation of a single command out of the source code and into a
standalone text file. It sets up the framework for converting such files
back into C source for use by the optget() function.

Related #507