Convert `alias` documentation to Sphinx/RST
This is a proof of concept change to implement the documentation using
Sphinx/RST markup. It includes documentation for the `alias` special
builtin and the `basename` command builtin.  The sole purpose is to
establish that using Sphinx/RST is viable by providing concrete examples
for examination.

I've deliberately chosen to include the documentation build artifacts
in the commit.  This means that for now the CI builds, and anyone else
building from source, doesn't need to have Sphinx installed to build the
documentation. This does mean that some care is needed to always run `make
man` and `make html` before committing any change to the documentation. TBD
is whether a trigger of some sort will be needed to ensure the two sets
of files are kept in sync.

Related #507