Replace commit 2d497fbc with a better approach
Having an independent set of vars contain pointers also stored in
shgd->bltin_nodes is sort of gross. Replace that approach with a
structure. That allow us to use a name instead of a possibly incorrect
array offset for quick access to a specific shell var. Now we only have
to worry about the number of vars changing and that is something we can
cheaply detect at run time.

Also implement a consistent naming scheme for the special symbol for each
shell var. The old scheme was very inconsistent. The new scheme has a
`VAR_` prefix for all of the vars and the suffix is the actual shell
var name.