Refactor b_print(), b_echo(), and b_printf()
The way in which these three commands were entangled made understanding
the behavior of each much more difficult. The code is also confusing for
other reasons such as aliasing variable `rflag` with `(struct print).raw`.
It also makes it hard to replace optget() with getopt_long(). So before
doing that refactor the code in question.