Move the logic for skipping tests
Move the logic for skipping tests from the meson.build file to the
run_test.sh script. This is for

a) consistency since some tests are already skipped in run_test.sh or in a
specific unit test, and

b) because it is better to have `meson test` show which tests are skipped
rather than during the `meson` config phase.

The reason the skip logic was originally in the meson.build config file
was that we didn't know about the magic `77` exit status that Meson
recognizes to mean the test was skipped rather than passed or failed.