Fix `cd` function and make it automatically available
User Andras Farkas <deepbluemistake@gmail.com> brought this issue to our
attention via a message on the ksh mailing list. There are two problems.
First, autoloaded functions `cd` and `mcd` aren't available until the
user uses the `dirs` command. So put each function definition in its own
autoloaded file. Second, the `_cd` function behaves incorrectly if you
invoke it with no arguments.

I'm ambivalent about the `_cd` behavior of always printing the selected
directory on stdout. Personally I think that is just noise. If the user
wants to be reminded what the CWD is they can put it in their prompt.
But I'm keeping that behavior for the moment to see how others feel and
because that is what the legacy implementation does which implies some
people think the behavior is useful.