Try to reduce false positive expect test failures
Since I modified the framework for running `expect` based tests to capture
more info when they timeout I've noticed that every single timeout happens
on the final [ctrl-D] to terminate the shell. Instead of terminating the
shell a new prompt appears. This is an attempt to make those failures
less likely.