Make `DPRINTF()` more robust
Once in a while I'll use `DPRINTF()` in a way that causes a SIGSEGV while
formatting the message. Usually because I'm trying to print a bogus pointer
via the `%s` formatting directive. This causes the code to catch that
failure and substitute a generic message rather than killing the process.