Fix printf %H
The man page says %H is meant to produce text valid for use in both HTML and XML.  The current behavior is both invalid for HTML and invalid for XML. &nbsp; is invalid in XML.  &apos; is invalid in HTML (not including HTML5)
Thus, I eliminate the invalid entities here, making it so that the output of printf %H is valid in both HTML and XML. This is backwards-incompatible as for ksh, but it has to be, since it was implemented wrong to begin with.  Fortunately, this doesn't affect compatibility with POSIX printf. ;D

%#H is already in use for making valid URIs...  So I don't know what to do to make a %something which will also include &apos; for XML.  What flag could we use for printf to print in a more XML-oriented way? i.e. to be the same as this functionality post-diff, but with &apos; added back.

I've tested the behavior of ksh itself, but haven't tested the tests.