Move the generated version from a .h to a .c file
Reinstate commit 4a438e51 that was reverted by commit 91b38069. That
commit reversion was a mistake because it appeared, based on the problem
description, that the change incorrectly constrained the valid version
strings.  Subsequent analysis of the problem showed that not to be the case
and the problem reporter was, in fact, attempting to use a non-semantic
version string.

The SH_VERSION preprocessor symbol has bothered me for a long time.
Especially as used in the context of `e_version[]`. This simplifies how
the version is injected by the build process into the code; notwithstanding
the large number of files touched by this change.