Fix cd function and enable it in unit tests
The old AST team implementation of an improved `cd` command has several
problems when enabled for every interactive shell. Not least of which
is that the approach of using an alias does not allow it to be, easily,
overriden by a user defined `cd` function. It also fails to preserve some
builtin `cd` command behavior. There may be other problems this doesn't
fix. This is an experiment to see if we can augment commands like `cd`
without having to write all behavior in the C language.  Something other
shells, like fish, have managed for more than a decade.

Fixes #1445