Deal with PATH being unset when running unit tests
This is nowhere close to optimal. It is a clumsy fix for when the
user's environment has $PATH unset when running unit tests. The primary
reason for this change is to replace the unnecessary invocation of
the external `basename` command with an equivalent ksh builtin when
running src/cmd/ksh93/data/config.ksh. A secondary reason is to work
around whatever is causing `$PATH` to be unset for the person who opened
issue #1453.

Long term we might want to augment the *config.ksh* script to also
explicitly deal with an empty/unset `$PATH`.

Fixes #1453