Eliminate bogus locale tests
Some platforms (e.g., OpenBSD) allow pretty much any locale string. And it
is not clear why it is the responsibility of ksh to issue a diagnostic if
an invalid locale is specified by the user. Or why we should be testing
for this situation. Even ksh93u+, which uses the AST locale subsystem,
doesn't produce a diagnostic message when doing `LC_CTYPE=xyz.UTF-8`
on OpenBSD. This is a follow up to commit 6b6146e7.